Fix date issues when fetching new recalls for notifications
- The server's date was in seconds while the one from the DB was in milliseconds.
- The user were not notified about recalls whose date was equal to the DB's most recent recall.
- Extracted method to `RecallRepository`

+++ app/src/main/java/com/sonphil/canadarecallsandsafetyalerts/api/recent/ApiRecall.kt
@@ -8,7 +8,7 @@ data class ApiRecall(
     @Json(name = "category")
     val category: List<Int>,
     @Json(name = "date_published")
-    val datePublished: Int?,
+    val datePublished: Long?,
     @Json(name = "recallId")
     val recallId: String,
     @Json(name = "title")
